requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
opencv-contrib-python==4.10.0.84
numpy==1.26.4
aiosqlite==0.20.0
pydantic==2.9.2
python-multipart==0.0.9
pillow==10.4.0